requirements.txt
python-dotenv==1.0.0
aiohttp==3.10.5
pyTelegramBotAPI
colorama